pyproject.toml
[project]
name = "cryptyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.10.2",
    "pydantic-settings>=2.6.1",
    "pygithub>=2.5.0",
    "python-dotenv>=1.0.1",
    "python-json-logger>=2.0.7",
    "reportlab>=4.2.5",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.1",
]
